refactor: move input validation to own file
See: BEDS-605
